requirements.txt
groq==0.1.0
azure-cognitiveservices-speech==1.31.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2
pytest==7.4.3
pytest-cov==4.1.0
black==23.9.1
mypy==1.5.1
pylint==3.0.2
python-json-logger==2.0.7
tenacity==8.2.3
types-requests==2.31.0.10
typing-extensions==4.8.0